backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
requests
whitenoise
gunicorn
Pillow
